requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
blas=1.0=mkl
brotlipy=0.7.0=py311h5eee18b_1002
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.05.30=h06a4308_0
certifi=2023.7.22=py311h06a4308_0
cffi=1.15.1=py311h5eee18b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
cryptography=41.0.2=py311h22a60cf_0
cudatoolkit=11.8.0=h6a678d5_0
cudnn=8.9.2.26=cuda11_0
cupti=11.8.0=he078b1a_0
filelock=3.9.0=py311h06a4308_0
freetype=2.12.1=h4a9f257_0
giflib=5.2.1=h5eee18b_3
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py311hc9b5ff0_0
idna=3.4=py311h06a4308_0
intel-openmp=2023.1.0=hdb19cb5_46305
jinja2=3.1.2=py311h06a4308_0
jpeg=9e=h5eee18b_1
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libdeflate=1.17=h5eee18b_0
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libpng=1.6.39=h5eee18b_0
libprotobuf=3.20.3=he621ea3_0
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.5.1=h6a678d5_0
libuuid=1.41.5=h5eee18b_0
libwebp=1.2.4=h11a3e52_1
libwebp-base=1.2.4=h5eee18b_1
lz4-c=1.9.4=h6a678d5_0
magma=2.7.1=h2c23e93_0
markupsafe=2.1.1=py311h5eee18b_0
mkl=2023.1.0=h213fc3f_46343
mkl-service=2.4.0=py311h5eee18b_1
mkl_fft=1.3.6=py311ha02d727_1
mkl_random=1.2.2=py311ha02d727_1
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.3.0=py311h06a4308_0
ncurses=6.4=h6a678d5_0
networkx=3.1=py311h06a4308_0
ninja=1.10.2=h06a4308_5
ninja-base=1.10.2=hd09550d_5
numpy=1.25.2=py311h08b1b3b_0
numpy-base=1.25.2=py311hf175353_0
openssl=3.0.10=h7f8727e_2
pillow=9.4.0=py311h6a678d5_0
pip=23.2.1=py311h06a4308_0
pycparser=2.21=pyhd3eb1b0_0
pyopenssl=23.2.0=py311h06a4308_0
pysocks=1.7.1=py311h06a4308_0
python=3.11.4=h955ad1f_0
pytorch=2.0.1=gpu_cuda118py311h7668aad_0
readline=8.2=h5eee18b_0
requests=2.31.0=py311h06a4308_0
setuptools=68.0.0=py311h06a4308_0
sqlite=3.41.2=h5eee18b_0
sympy=1.11.1=py311h06a4308_0
tbb=2021.8.0=hdb19cb5_0
tk=8.6.12=h1ccaba5_0
torchvision=0.15.2=cuda118py311h4cc2eb7_0
typing-extensions=4.7.1=py311h06a4308_0
typing_extensions=4.7.1=py311h06a4308_0
tzdata=2023c=h04d1e81_0
urllib3=1.26.16=py311h06a4308_0
wheel=0.38.4=py311h06a4308_0
xz=5.4.2=h5eee18b_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
